(PC-7408) move Section and SectionRow components to ui/
