+++ src/ui/components/Section.tsx
@@ -4,14 +4,14 @@ import styled from 'styled-components/native'
 
 import { getSpacing, Typo } from 'ui/theme'
 
-import { Separator } from './reusables'
+import { Separator } from './Separator'
 
-interface ProfileSectionProps {
+interface SectionProps {
   title?: string
   style?: StyleProp<ViewStyle>
 }
 
-export function ProfileSection(props: PropsWithChildren<ProfileSectionProps>) {
+export function Section(props: PropsWithChildren<SectionProps>) {
   return (
     <Container style={props.style}>
       {props.title && <Title>{props.title}</Title>}
