+++ src/ui/components/Separator.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components/native'
+
+import { ColorsEnum } from 'ui/theme'
+
+export const Separator = styled.View({
+  width: '100%',
+  height: 1,
+  backgroundColor: ColorsEnum.GREY_LIGHT,
+})
